+++ CHANGELOG.md
@@ -116,5 +116,5 @@
 
 ### Changed
 
-- Added experimental [`zip::unstable::write::FileOptions::with_deprecated_encryption`] API to enable encrypting files 
-  with PKWARE encryption.
+- Added experimental [`zip_next::unstable::write::FileOptions::with_deprecated_encryption`] API to enable encrypting
+  files with PKWARE encryption.